Fix MSVC narrowing warnings in newer MSVC versions.
Bug: dawn:824
Reviewed-on: https://dawn-review.googlesource.com/c/tint/+/75073
Auto-Submit: Corentin Wallez <cwallez@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Ben Clayton <bclayton@google.com>

+++ src/resolver/resolver.cc
@@ -1565,7 +1565,7 @@ sem::Call* Resolver::TypeConstructor(
                     [&](const sem::Type* t, size_t i) -> const sem::Parameter* {
                       return builder_->create<sem::Parameter>(
                           nullptr,                   // declaration
-                          i,                         // index
+                          static_cast<uint32_t>(i),  // index
                           t->UnwrapRef(),            // type
                           ast::StorageClass::kNone,  // storage_class
                           ast::Access::kUndefined);  // access
